additional Executor changes
